fixed result() of PStep in non-blocking mode fixed getRunningPandaJobs in psequencer

+++ pandatools/PSeqUtils.py
@@ -26,7 +26,7 @@ class PandaJobFactory:
         # instantiate PStep
         pStep = PStep.PStep(stepName,'',self.fetFactory,self.sn,self.verbose)
         # set JobID and Panda mark
-        pStep.JobID   = jobID
+        pStep.JobID   = long(jobID)
         pStep.isPanda = True
         # return
         return pStep
